Config // moved static factory methods into "ConfigBuilder"-class ConfigBuilder // added support for base64 encoded args
